Add files via upload
System health monitoring is a set of activities undertaken to maintain a system in operable condition that may be limited to an observation of current system states, with maintenance and repair being prompted by these observations. In an operating room, you will typically find some form of thermometer to track a patient's temperature, a pulse oximeter to measure oxygen levels, a capnography machine for monitoring CO2 levels, and a sphygmomanometer to measure blood pressure.
The objective of developing monitoring systems is to reduce healthcare costs by reducing physician office visits, hospitalizations, and diagnostic testing procedure. The GSM technology helps the server to update the patient data on the website.